+++ docs/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.5.0 (2022-01-06)
 
 - feat: support creating and calling custom functions with [go plugin](https://pkg.go.dev/plugin)
+- feat: install hrp with one shell command
+- feat: report GA event for loading go plugin
 
 ## v0.4.0 (2022-01-05)